requirements.txt
Django==1.9.7
django-appconf==1.0.2
django-ckeditor==5.0.3
django-imagekit==3.3
djangorestframework==3.3.3
pilkit==1.1.13
Pillow==3.3.0
six==1.10.0
Unidecode==0.4.19